long long value ITOA && INTLEN

+++ ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: flbeaumo <flbeaumo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/24 12:22:15 by flbeaumo          #+#    #+#             */
-/*   Updated: 2019/05/24 13:58:31 by flbeaumo         ###   ########.fr       */
+/*   Updated: 2019/05/25 10:55:18 by flbeaumo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int     ft_intlen(long nb)
+int     ft_intlen(long long nb)
 {
     int count;
